backend/requirements.txt
ephem==4.1.4
pyswisseph==2.10.3.2
fastapi==0.104.1
uvicorn==0.24.0
python-dateutil==2.8.2
pydantic==2.5.0
Flask==3.0.0
python-dotenv==0.21.0
Flask-Cors==4.0.1
